refactor : model 네이밍 domain으로 변경
