+++ src/main/java/racingcar/domain/Cars.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import racingcar.dto.CarDto;
 import racingcar.dto.GamePlayResponseDto;
