+++ src/main/java/racingcar/domain/Name.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import racingcar.exception.BlankNameException;
 import racingcar.exception.WrongNameLengthException;
