+++ src/main/java/racingcar/domain/Round.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import racingcar.exception.WrongRoundException;
